+++ src/extractors/index.ts
@@ -1,6 +1,6 @@
-export { isAxiosCall, extractAxiosCall } from "./axiosExtractor";
-export { isFetchCall, extractFetchCall } from "./fetchExtractor";
-export { isGotCall, extractGotCall } from "./gotExtractor";
-export { isRequestCall, extractRequestCall } from "./requestExtractor";
-export { isSuperagentCall, extractSuperagentCall } from "./superagentExtractor";
-export { isNodeFetchCall, extractNodeFetchCall } from "./nodeFetchExtractor";
+export { isAxiosCall, extractAxiosCall } from './axiosExtractor';
+export { isFetchCall, extractFetchCall } from './fetchExtractor';
+export { isGotCall, extractGotCall } from './gotExtractor';
+export { isRequestCall, extractRequestCall } from './requestExtractor';
+export { isSuperagentCall, extractSuperagentCall } from './superagentExtractor';
+export { isNodeFetchCall, extractNodeFetchCall } from './nodeFetchExtractor';