+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 
 All notable changes to this project will be documented in this file.
 
+## v1.0.1-beta.1 (2024-09-01)
+
+### added
+
+- Husky pre commit added.
+- License added.
+
+### changed
+
+- non useful comments removed.
+- eslint config has been changes.
+- verbose logs removed from the tests.
+
 ## v1.0.0-beta.1 (2024-08-31)
 
 First release